Update version 0.5
