+++ screens.h
@@ -1,0 +1,12 @@
+#ifndef screens_H
+#define screens_H
+#include <stdio.h>
+#include <stdlib.h>
+char main_menu_screen(void);
+void about_screen(void);
+void team_screen(void);
+char login_screen(void);
+void registration_screen(void);
+void login_client(void);
+void login_admin(void);
+#endif
